OptimPhase2-package(OptimPhase2)	R Documentation

Two-stage design for comparison of a time to event endpoint to a known control with potential stopping for futility after the first stage.

Description

OptimDes is the primary function in package OptimPhase2. Detailed documentation of the design is in method.pdf, which can be found under "directory" on the package HTML help page. The user supplies the projected accrual distribution and distributions for time to event under the null and alternative hypotheses. OptimDes returns the total sample size, the interim time and sample size, and boundaries for decision rules.

plot.OptimDes can be applied to the output of OptimDes to evaluate other two-stage designs that may achieve near optimal results for both of the implemented optimality criteria. print.OptimDes produces a summary of the design with a comparison to a standard single-stage design (which can also be obtained from FixDes) Design parameters for nearly optimal designs can be evaluated with np.OptimDes. Statistical tests for a two-stage design are implemented in Test2stage.

The function SimDes produces simulations under the assumed design by default to check the accuracy of asymptotic approximations. By supplying alterative accrual and survival distributions, the performance of the design and estimation can be checked for robustness.

The function weibPmatch can be used to find parameters of a Weibull distribution that match a specified event rate at a specified time. Function weibull.plot displays a null-alternative pair of hypothesized Weibull survival functions.

Details

Package: 	 OptimPhase2

Type: 	 Package

Version: 	 1.2

Date: 	 2006-10-26

License: 	 GPL version 2 or later

Author(s)

Bo Huang <bo.huang@pfizer.com> and Neal Thomas <neal.thomas@pfizer.com>

See Also

Package Survival

[Package OptimPhase2 version 1.5 Index]