Test2stage {OptimPhase2}	R Documentation

Statistical test for two-stage designs from function OptimDes

Description

This function performs the hypothesis tests for the two-stage designs with time to event endpoint from OptimDes.

Usage


Test2stage(Y1, T1, Y2 = NULL, T2 = NULL, p0, x, C1, C2, t1, MTSL =
NULL, printTest = TRUE, cen1=rep(1,length(T1)),cen2=rep(1,length(T2)))

Arguments

Y1 	 A vector containing the study start times (measured from the beginning of the study) of first-stage patients.

T1 	 A vector containing the event times of first-stage patients.

Y2 	 A vector containing the study start times (measured from the beginning of the study) of second-stage patients.

T2 	 A vector containing the event times of second-stage patients.

p0 	 The event rate under the null hypothesis.

x 	 Survival time of interest (e.g., 1 year).

C1 	 The two-stage study is terminated for futility after the first stage if the Z-statistic is < C1

C2 	 Reject the null hypothesis for the one-sided alternative hypothesis at the second stage if the Z-statistic>C2.

t1 	 Time for the interim analysis (actual time, which may differ from the planned time).

MTSL 	 Maximum total study length, (actual, not planned)

printTest 	 If TRUE (default), the result of the test and the interim decision is printed.

cen1 	 The times in T1 are regarded as events unless they are set to censored by setting the corresponding value in cen1 to zero.

cen2 	 The times in T2 are regarded as events unless they are set to censored by setting the corresponding value in cen1 to zero.

Details

The hypothesis tests are performed in two stages as described in Case and Morgan (2003).

Stage 1. Accrue patients between time 0 and time t1. Each patient will be followed until failure, or for x years or until time t1, whichever is less. Calculate the normalized interim test statistic Z1. If Z1<C1, stop the study for futility; otherwise, continue to the next stage.

Stage 2. Accrue patients between t1 and MDA. Follow all patients until failure or for x years, then calculate the normalized final test statistic Z2, and reject H0 if Z2>C2.

The test statistic is based on the Nelson-Aalen estimator of the cumulative hazard function.

Value

A vector containing results for the interim analysis if T2 and Y2 are NULL, and the final analysis if T2 or Y2 are non-NULL:

z 	 The test statistic

se 	 Standard error of the cummulative hazard (not log cummulative hazard) at time x.

cumL 	 Cummulative hazard estimator at time x.

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Case M. D. and Morgan T. M. (2003) Design of Phase II cancer trials evaluating survival probabilities. BMC Medical Research Methodology, 3, 7.

Lin D. Y., Shen L., Ying Z. and Breslow N. E. (1996) Group seqential designs for monitoring survival probabilities. Biometrics, 52, 1033–1042.

Simon R. (1989) Optimal two-stage designs for phase II clinical trials. Controlled Clinical Trials, 10, 1–10.

See Also

OptimDes, SimDes

Examples


B.init <- c(1, 2, 3, 4, 5)
m.init <- c(15, 20, 25, 20, 15)
alpha <- 0.05
beta <- 0.1
param <- c(1, 1.09, 2, 1.40)
x <- 1

# H0: S0=0.40 H1: S1=0.60

shape0 <- param[1]
scale0 <- param[2]
shape1 <- param[3]
scale1 <- param[4]

object1 <- OptimDes(B.init,m.init,alpha,beta,param,x,target="EDA")
n <- object1$n[2]
t1 <- object1$stageTime[1]
C1 <- object1$boundary[1]
C2 <- object1$boundary[2]
b <- length(B.init)
l <- rank(c(cumsum(m.init),n),ties.method="min")[b+1]
mda <- ifelse(l>1,B.init[l-1]+(B.init[l]-B.init[l-1])*(n-sum(m.init[1:(l-1)]))/m.init[l],B.init[l]*(n/m.init[l]))

### set up values to create a stepwise uniform distribution for accrual
B <- B.init[1:l]
B[l] <- mda
xv <- c(0,B)
M <- m.init[1:l]
M[l] <- ifelse(l>1,n-sum(m.init[1:(l-1)]),n)
yv <- c(0,M/(diff(xv)*n),0)

# density function of accrual 
dens.Y <- stepfun(xv,yv,f=1,right=TRUE)
# pool of time points to be simulated from
t.Y <- seq(0,mda,by=0.01)

# simulate study times of length n
sample.Y <- sample(t.Y,n,replace=TRUE,prob=dens.Y(t.Y))

# simulate failure times of length n under the alternative hypothesis
sample.T <- rweibull(n,shape=shape1,scale=scale1)

Y1 <- sample.Y[sample.Y<=t1]
T1 <- sample.T[sample.Y<=t1]
Y2 <- sample.Y[sample.Y>t1]
T2 <- sample.T[sample.Y>t1]

# event rate under null hypothesis
p0<-pweibull(x,shape=shape0,scale=scale0)

# interim analysis
Test2stage(Y1, T1, Y2 = NULL, T2 = NULL, p0, x, C1, C2, t1, MTSL = NULL)

# final analysis if the study continues
Test2stage(Y1, T1, Y2, T2, p0, x, C1, C2, t1, mda+x)

# simulate failure times of length n under the null hypothesis
sample.T <- rweibull(n,shape=shape0,scale=scale0)

Y1 <- sample.Y[sample.Y<=t1]
T1 <- sample.T[sample.Y<=t1]
Y2 <- sample.Y[sample.Y>t1]
T2 <- sample.T[sample.Y>t1]

# interim analysis
Test2stage(Y1, T1, Y2 = NULL, T2 = NULL, p0, x, C1, C2, t1, MTSL = NULL)

# final analysis if the study continues
Test2stage(Y1, T1, Y2, T2, p0, x, C1, C2, t1, mda+x)


[Package OptimPhase2 Index]