plot.OptimDes {OptimPhase2}	R Documentation

Plot efficiency of optimal two-stage designs as a function of total sample size

Description

Output from function OptimDes is used to display the ETSL and EDA for a two-stage design relative to a single-stage design as a function of the combined stage 1 and 2 sample size.

Usage


## S3 method for class 'OptimDes':
plot(x, xscale = "t", l.type = 1:4, l.col =
c("blue", "green", "red"), CMadj=F, ...)

Arguments

x 	 Output from function OptimDes.

xscale 	 Scale of the x-axis. "t" for combined stage 1 and 2 study length. "n" for combined sample size. Default is t.

l.type 	 Line types for the plot. Default is 1-4.

l.col 	 Line colors for the plot. Default is "blue" for ETSL, "green" for EDA and "red" for t1.

CMadj 	 If true, the sample sizes and times are adjusted by the ratio of the exact binomial to asymptotic normal sample size for the single stage design, as in Case and Morgan 2003. Proportional adjustment of times and samplesizes are made even if the accrual rates are not constant. Default is false.

... 	 Additional graphical parameters passed to function plot.

Details

The plot displays the tradeoff between ETSL and EDA as a function of the combined sample size. The robustness of the optimal two-stage design to deviations from the target sample size can be explored. The plots often suggest compromise designs achieving near-optimal results for both EDA and ETSL. Test boundary values (C1, C2), and numerical values of other design parameters, can be obtained for a design selected from the plots using function np.OptimDes.

The plot also includes the time of the interim analysis (t1) as a ratio to the time for a corresonding single-stage analysis.

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Case M. D. and Morgan T. M. (2003) Design of Phase II cancer trials evaluating survival probabilities. BMC Medical Research Methodology, 3, 7.

Lin D. Y., Shen L., Ying Z. and Breslow N. E. (1996) Group seqential designs for monitoring survival probabilities. Biometrics, 52, 1033–1042.

Simon R. (1989) Optimal two-stage designs for phase II clinical trials. Controlled Clinical Trials, 10, 1–10.

See Also

print.OptimDes, OptimDes

[Package OptimPhase2 Index]