print.OptimDes {OptimPhase2}	R Documentation

Print Objects Output By OptimDes

Description

Print an object output by OptimDes.

Usage


## S3 method for class 'OptimDes':
print(x, dig = 3, all=FALSE, condPow=F,
                     CMadj=F, ...)

Arguments

x 	 Object output by OptimDes.

dig 	 Number of digits printed.

all 	 If TRUE, results are printed for all total sample sizes evaluated by OptimDes.

condPow 	 Conditional power at the test boundary C1 under H1 is reported when condPow=T

CMadj 	 If true, the sample sizes and times are adjusted by the ratio of the exact binomial to asymptotic normal sample size for the single stage design, as in Case and Morgan 2003. Proportional adjustment of times and sample sizes are made even if the accrual rates are not constant. Default is false.

... 	 Optional print arguments, see printt.default.

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Case M. D. and Morgan T. M. (2003) Design of Phase II cancer trials evaluating survival probabilities. BMC Medical Research Methodology, 3, 7.

Lin D. Y., Shen L., Ying Z. and Breslow N. E. (1996) Group seqential designs for monitoring survival probabilities. Biometrics, 52, 1033–1042.

Simon R. (1989) Optimal two-stage designs for phase II clinical trials. Controlled Clinical Trials, 10, 1–10.

See Also

plot.OptimDes, OptimDes

[Package OptimPhase2 Index]