weibull.plot {OptimPhase2}	R Documentation

Plot Weibull Survival Curves

Description

Plot Weibull survival curves with differences at a target time highlighted. Designed for use with the param values input to function OptimDes.

Usage


weibull.plot(param, x, l.type = 1:3, l.col = c("blue", "red"), ...)

Arguments

param 	 A four-element vector containing: the shape parameter of the Weibull distribution under the null hypothesis, the scale parameter of the Weibull distribution under the null hypothesis, the shape parameter of the Weibull distribution under the alternative hypothesis, the scale parameter of the Weibull distribution under the alternative hypothesis.

x 	 Survival time of interest (e.g., 1 year).

l.type 	 Line types for the plot. Default is 1-3.

l.col 	 Line colors for the plot. Default is "blue" for the null survival curve, "red" for the alternative survival curve.

... 	 Further graphical arguments, see plot.default.

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.

See Also

dweibull, OptimDes, weibPmatch

Examples


param <- c(1, 1.09, 2, 1.40)
x <- 1

weibull.plot(param,x)


[Package OptimPhase2 Index]