FixDes(OptimPhase2)	R Documentation

Construct a Single-stage Design for a Time to Event Endpoint Versus a Known Control Standard

Description

Find the sample size, duration of accrual, and test boundary for a single-stage design with an event-free endpoint versus a known control standard. Testing is one-sided based on the Kaplan-Meier estimator evaluated at a user-specified time point.

Usage


FixDes(B.init, m.init, alpha, beta, param, x)

Arguments

B.init 	 A vector of user-specified time points (B1, ..., Bb) that determine a set of time intervals with uniform accrual.

m.init 	 The projected number of patients that can be accrued within the time intervals determined by B.init.

alpha 	 Type I error.

beta 	 Type II error.

param 	 Events should be defined as poor outcomes. Computations and reporting are based on the proportion without an event at a pre-specified time, x. For constructing an optimal design, complete event-free distributions at all times must be specified for the control condition (Null), and for the alternative "effective" treatment. Weibull distributions are currently implemented. param is a vector of length 4: (shape null, scale null, shape alternative, scale alternative). The R parameterization of the Weibull distribution is used.

x 	 Pre-specified time for the event-free endpoint (e.g., 1 year).

Details

Estimation is based on the Kaplan-Meier or Nelson-Aalen estimators evaluated at a target time (e.g., 1 year). The event-free rates at the target time are computed from Weibull distributions assumed for the treated and (assumed known) control distributions, as is done in function OptimDes. The design depends only on the event-free rates at the target time (except for small changes due to rounding with different survivor functions). The duration of accrual depends on the projected maximum accrual rates.

Value

A list with components:

n0 	 Fixed design sample size

DA 	 Duration of accrual

SL 	 Total study length (time, DA+x).

n0E 	 n0 based on exact binomial test

DAE 	 DA based on exact binomial test

SLE 	 SL based on exact binomial test

C 	 Rejection cutpoint for the test statistic.

Note

The single-stage sample size is used as the starting value for evaluating the optimal n for a two-stage design in OptimDes.

Author(s)

Bo Huang <bo.huang@pfizer.com> and Neal Thomas <neal.thomas@pfizer.com>

References

Case M. D. and Morgan T. M. (2003) Design of Phase II cancer trials evaluating survival probabilities. BMC Medical Research Methodology, 3, 7.

Lin D. Y., Shen L., Ying Z. and Breslow N. E. (1996) Group seqential designs for monitoring survival probabilities. Biometrics, 52, 1033–1042.

Simon R. (1989) Optimal two-stage designs for phase II clinical trials. Controlled Clinical Trials, 10, 1–10.

See Also

OptimDes, Test2stage, SimDes

Examples


B.init <- c(1, 2, 3, 4, 5)
m.init <- c(15, 20, 25, 20, 15)
alpha <- 0.05
beta <- 0.1
param <- c(1, 1.09, 2, 1.40)
x <- 1

# H0: S0=0.40 H1: S1=0.60

FixDes(B.init, m.init, alpha, beta, param, x)


[Package OptimPhase2 version 1.8 Index]