OptimDes {OptimPhase2}	R Documentation

Construct an Optimal Two-stage Design for a Time to Event Endpoint Versus a Known Control Standard

Description

Construct a two-stage design for a time to event endpoint with possible stopping for futility after the end of Stage 1. The design minimizes either the expected duration of accrual (EDA) or the expected total study length (ETSL). Requires 2-3 minutes on a 2GHz laptop.

Usage


OptimDes(
B.init, m.init, alpha, beta, param, x, target = c("EDA", "ETSL"),
recover=TRUE, control = OptimDesControl(),...)


Arguments

B.init 	 A vector of user-specified time points (B1, ..., Bb) that determine a set of time intervals with uniform accrual.

m.init 	 The projected number of patients that can be accrued within the time intervals determined by B.init. A large number of potential patients results in long execution times for OptimDes, so unrealistically large values should not be entered.

alpha 	 Type I error.

beta 	 Type II error.

param 	 A four-element vector containing: the shape parameter of the Weibull distribution under the null hypothesis, the scale parameter of the Weibull distribution under the null hypothesis, the shape parameter of the Weibull distribution under the alternative hypothesis, the scale parameter of the Weibull distribution under the alternative hypothesis. The R parameterization of the Weibull distribution is used.

x 	 Survival time of interest (e.g., 1 year).

target 	 When the expected duration of accrual (EDA) is to be minimized, target="EDA". When the expected total study length is to be minimized, target="ETSL".

recover 	 The Simon Two-Stage design recovers alpha from the Type I errors that would have occured if the study had not been stopped early for futility. This recovery allows for a slightly lower boundary for the final test, thereby increasing power. The default with partial data at the interim analysis is to recover the alpha as with the standard Simon design. When recover=F, the optimal design is found assuming that the boundary for the final test will not be adjusted downward ignoring this small potential gain to ensure protection of the alpha level even when the accrual rates and survivor functions are mis-specified.

control 	 An optional list of control settings. See OptimDesControl for the parameters that can be set and their default values.

... 	 No additional optional parameters are currently implemented

Details

OptimDes finds a two-stage design for a time to event endpoint with potential stopping for futility after the first stage. It implements the Case and Morgan (2003) generalizaton of the Simon (1989) two-stage design for comparing a treatment to a known standard. The design minimizes either the expected duration of accrual (EDA) or the expected total study length (ETSL).

The design calculations assume Weibull distributions for the time to event endpoint in the treated group, and for the (assumed known, "Null") control distribution. Estimation is based on the Kaplan-Meier or Nelson-Aalen estimators evaluated at a target time (e.g., 1 year). The full treated and control distributions and the accrual distribution affect power (and alpha level in some settings), see Case and Morgan (2003)).

Accrual rates are specified by the user. These rates can differ across time intervals specified by the user (this generalizes the results in Case and Morgan).

A detailed description of the algorithm is in the file method.pdf in the /doc subdirectory of the OptimPhase2 package. It can be accessed from the HTML help page for the package.

Value

A list with components:

target 	 The optimizaton target ("EDA" or "ETSL")

test 	 A vector giving the type I error alpha, type II error beta, recover=T/F, Weibull parameters param and survival time of interest x.

accrual 	 A list containing the input vectors B.init and m.init

result 	 A three-element vector containing the expected duration of accrual (EDA), the expected total study length (ETSL), and the expected sample size of the optimal design (ES).

n 	 A two-element vector containing the sample size for the interim analysis and the sample size if both stages are completed.

stageTime 	 A two-element vector giving the times for the interim and final analysis.

boundary 	 A two-element vector giving the rejection cutpoint C1 for the interim test statistic (see Test2stage), and C2 for the final test statistic.

se 	 A vector of length 4 with the asymptotic standard errors at the iterim and final analysis under the null hypothesis, followed by the corresponding SEs under the alternative hypothesis. These SEs must be divided by the sqrt(sample size).

u 	 Mean of Z2 under H1. See detailed description. Used to compute conditional power

exposure 	 The expected total exposure of patients at the time of the planned interim analysis. The exposure is truncated by both the interim analysis time and the target surival time (i.e., no exposure after x). Exposure is a vector of length 2. The first valuse is the expected exposure without the Case-Morgan finite sample adjustment to sample sizes. The second value is the expected exposure with theadjustment applied.

all.info 	 A data frame containing the results for all of the evaluated sample sizes.

single.stage 	 A six-element vector giving the sample size fix.n, duration of accrual DA, study length SL, and corresponding values based on the exact binomial test for a single-stage design with the design distributional assumptions.

Note

The algorithm will search for the optimal n between the sample size for a single-stage design and the user specified maximum sample size sum(m.init).

When the length of B.init or m.init is 1, the accrual rate is constant as in Lin et al. (1996), Case and Morgan (2003).

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Case M. D. and Morgan T. M. (2003) Design of Phase II cancer trials evaluating survival probabilities. BMC Medical Research Methodology, 3, 7.

Lin D. Y., Shen L., Ying Z. and Breslow N. E. (1996) Group seqential designs for monitoring survival probabilities. Biometrics, 52, 1033–1042.

Simon R. (1989) Optimal two-stage designs for phase II clinical trials. Controlled Clinical Trials, 10, 1–10.

See Also

np.OptimDes, print.OptimDes, plot.OptimDes, weibull.plot

Examples


B.init <- c(1, 2, 3, 4, 5)
m.init <- c(15, 20, 25, 20, 15)
alpha <- 0.05
beta <- 0.1
param <- c(1, 1.09, 2, 1.40)
x <- 1

# H0: S0=0.40 H1: S1=0.60

object1 <- OptimDes(B.init,m.init,alpha,beta,param,x,target="EDA")
print(object1)

object2 <- OptimDes(B.init,m.init,alpha,beta,param,x,target="ETSL")
plot(object2)

m.init <- c(25,20,15,20,25)
object3 <- OptimDes(B.init,m.init,alpha,beta,param,x,target="ETSL")


[Package OptimPhase2 Index]