OptimDesControl {OptimPhase2}	R Documentation

Set parameters controlling numerical methods for OptimDes and np.OptimDes

Description

Set parameters controlling numerical methods for OptimDes and np.OptimDes

Usage


OptimDesControl(trace=FALSE,tol=0.01,conv=1e-6,rho.int=c(0,1),aboveMin=1.05)

Arguments

trace 	 A logical value indicating if a trace of the iteration progress should be printed. Default is FALSE. If TRUE the sample size n, the corresponding optimal correlation rho and minimized EDA(ETSL) are printed at the conclusion of each iteration.

tol 	 tol is the desired accuracy for optimize. Default is 0.01.

conv 	 The desired convergence accuracy for C1 and C2. Default is 1e-6.

rho.int 	 A vector containing the end-points of the interval to be searched for the optimal rho. Default is (0, 1).

aboveMin 	 The minimization method searches by increasing order from the single-stage sample size until the criteria exceeds the current minimum by a multiplicative factor of aboveMin. The search will also terminate if sum(m.init) is reached.

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Case M. D. and Morgan T. M. (2003) Design of Phase II cancer trials evaluating survival probabilities. BMC Medical Research Methodology, 3, 7.

Lin D. Y., Shen L., Ying Z. and Breslow N. E. (1996) Group seqential designs for monitoring survival probabilities. Biometrics, 52, 1033–1042.

Simon R. (1989) Optimal two-stage designs for phase II clinical trials. Controlled Clinical Trials, 10, 1–10.

See Also

OptimDes, np.OptimDes

[Package OptimPhase2 Index]