weibPmatch {OptimPhase2}	R Documentation

Compute the shape or scale parameter for a Weibull distribution so it has a specified event rate at a specified time.

Description

Determine the shape or scale parameter of a Weibull distribution so it has event rate P0 at time x. If the shape is specified, the scale parameter is computed, and if the scale is specified, the shape parameter is computed.

Usage


weibPmatch(x, p0, shape, scale)

Arguments

x 	 Time for targetted event rate

p0 	 Event rate at time x (1- survival rate).

shape 	 If specified, the necessary scale parameter is computed

scale 	 If specified, the necessary shape parameter is computed

Details

The time and event rate must be supplied. Either the shape or scale parameter must also be specified, but not both. The R parameterization of the Weibull distribution is used.

Value

A single numerical value is returned, either the shape or scale parameter, depending on which is specified by the user.

Author(s)

Bo Huang <huang@stat.wisc.edu> and Neal Thomas

References

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.

See Also

weibull.plot,pweibull, OptimDes

Examples



library(OptimPhase2)
param <- c(1, 1.09, 2, 1.40)
x<-1

p0<-pweibull(x,param[1],param[2])
p1<-pweibull(x,param[3],param[4])

weibull.plot(param,x)

### equivalent to simple call
paramNew<-c(param[1], weibPmatch(x,p0,param[1]), param[3], weibPmatch(x,p1,param[3]))
weibull.plot(paramNew, x)

### null curve with different shape
paramNew<-c(3.0, weibPmatch(x,p0,3.0), param[3], weibPmatch(x,p1,param[3]))
weibull.plot(paramNew, x)

### alternative curve with different scale
paramNew<-c(param[1], param[2], weibPmatch(x,p1,scale=2), 2)
weibull.plot(paramNew, x)


[Package OptimPhase2 Index]