weibull.plot(OptimPhase2)	R Documentation

Plot Weibull Survival Curves

Description

Plot Weibull survival curves with differences at a target time highlighted. Designed for use with the param values input to function OptimDes.

Usage


weibull.plot(param, x, l.type = 1:3, l.col = c("blue", "red"), ...)

Arguments

param 	 Events should be defined as poor outcomes. Computations and and reporting are based on the proportion without an event at a pre-specified time, x. For constructing an optimal design, complete event-free distributions at all times must be specified for the control condition (Null), and for the alternative "effective" treatment. Weibull distributions are currently implemented. param is a vector of length 4: (shape null, scale null, shape alternative, scale alternative). The R parameterization of the Weibull distribution is used.

x 	 Pre-specified time for the event-free endpoint (e.g., 1 year).

l.type 	 Line types for the plot. Default is 1-3.

l.col 	 Line colors for the plot. Default is "blue" for the null survival curve, "red" for the alternative survival curve.

... 	 Further graphical arguments, see plot.default.

Author(s)

Bo Huang <bo.huang@pfizer.com> and Neal Thomas <neal.thomas@pfizer.com>

References

Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, volume 1, chapter 21. Wiley, New York.

See Also

dweibull, OptimDes, weibPmatch

Examples


param <- c(1, 1.09, 2, 1.40)
x <- 1

weibull.plot(param,x)


[Package OptimPhase2 version 1.6 Index]